Split terminal output scheduler pipeline
